working version: fixed date format; temporary fix on coordinates (was floats, now int temporarily) due to https://github.com/CyberCRI/RedMetrics/issues/10

+++ RedMetricsDummy/Assets/Redmetrics/example/SerializationTest.cs
@@ -12,7 +12,7 @@ public class SerializationTest : MonoBehaviour {
 		customData.Add("customKey1", "customValue1");
 		customData.Add("customKey2", "customValue2");
 
-		float[] coordinates = { 1, 2, 3 };
+		int[] coordinates = { 1, 2, 3 };
 
 		TrackingEventData data = new TrackingEventDataWithoutIDs(
 			TrackingEvent.START,
